+++ src/core/components/RendererGroup.cs
@@ -8,7 +8,7 @@ public class RendererGroup : Component
 {
     public override int Order => ComponentOrder.Graphics;
 
-    public short Layer { get; set; }
+    public sbyte Layer { get; set; }
     public SpriteEffects Effects { get; set; }
 
     private SortedDictionary<int, HashSet<Renderer>> RenderersByLayer { get; }
