+++ src/core/components/SpriteRenderer.cs
@@ -23,7 +23,7 @@ public class SpriteRenderer(Entity entity) : Renderer(entity)
 
     public override Vector2 Size => Sprite == null ? Vector2.Zero : Sprite.Rectangle.Size.ToVector2();
 
-    public override void OnDraw(Graphics graphics, short layer, Vector2 sortingOrigin = default)
+    public override void OnDraw(Graphics graphics, sbyte layer, Vector2 sortingOrigin = default)
     {        
         Engine.Graphics.Draw(Entity.Node, new SpriteDrawer(Sprite, Effects), Color, layer, sortingOrigin);
     }
