+++ src/core/components/TextRenderer.cs
@@ -50,7 +50,7 @@ public class TextRenderer : Renderer, ISizable
         Font = Engine.Content.Load<SpriteFont>("Arial");
     }
 
-    public override void OnDraw(Graphics graphics, short layer, Vector2 sortingOrigin = default)
+    public override void OnDraw(Graphics graphics, sbyte layer, Vector2 sortingOrigin = default)
     {
         graphics.Draw(Entity.Node, new TextDrawer(Text, Font, Effects), Color, layer, sortingOrigin);
     }
