+++ src/core/components/TilemapRenderer.cs
@@ -22,7 +22,7 @@ public class TilemapRenderer : Renderer
         Tilemap ??= Entity.GetComponent<Tilemap>();
     }
 
-    public override void OnDraw(Graphics graphics, short layer, Vector2 sortingOrigin = default)
+    public override void OnDraw(Graphics graphics, sbyte layer, Vector2 sortingOrigin = default)
     {
         if (Tilemap != null)
         {
